Consul: Split up CATS and add releasable job
Separate teardown from CATS job so that we can run
CATS multiple times more quickly and get more feedback
when there are flaky tests.

Split releasable job from create-final-release so that
it is more easy to tell which commit of consul-release
will be released when the create-final-release job is
triggered or unpaused.

Signed-off-by: Desmond Pompa Alarcon Rawls <drawls@pivotal.io>